Put AyaArticle and AyaArticles in separate folders with .scss.  Styles article page.
