+++ src/containers/Aya/AyaArticle/AyaArticle.js
@@ -8,7 +8,7 @@ import Article from 'helpers/Article';
 import {BreadcrumbContainer} from 'components';
 import {Breadcrumb} from 'react-bootstrap';
 import { LinkContainer } from 'react-router-bootstrap';
-import heroStyles from './heroPanel.scss';
+import heroStyles from '../heroPanel.scss';
 
 @connect(
   (store) => {
@@ -32,13 +32,14 @@ export default class ArticleContainer extends Component {
 
   render() {
     const {articleHelper, isFetching} = this.props;
+    const styles = require('./AyaArticle.scss');
     return (
       <ArticleComponent
           article={articleHelper}
           isFetching={isFetching}
           heroStyles={heroStyles}
           >
-         <BreadcrumbContainer className="breadcrumb-wrapper--aya">
+         <BreadcrumbContainer className={`${styles.breadcrumbwrapper} breadcrumb-wrapper--aya`}>
           <LinkContainer to={'/aya/article/list'}>
             <Breadcrumb.Item>Articles</Breadcrumb.Item>
           </LinkContainer>
