+++ src/containers/Aya/AyaArticleList/AyaArticles.scss
@@ -1,0 +1,8 @@
+@import 'src/theme/variables';
+
+.promoUnit a {
+    text-decoration: none;
+}
+.iconWrapper {
+  background-color: $aya-green;
+}
