+++ src/containers/index.js
@@ -24,7 +24,7 @@ export UpdateImage from './Images/UpdateImage/UpdateImage';
 export AddImage from './Images/AddImage/AddImage';
 export DeleteImage from './Images/DeleteImage/DeleteImage';
 export Article from './Articles/Article/Article';
-export AyaArticle from './Aya/Article';
+export AyaArticle from './Aya/AyaArticle/AyaArticle';
 export ArticleList from './Articles/ArticleList/ArticleList';
 export InspirationList from './Inspiration/InspirationList/InspirationList';
 export ViewInspiration from './Inspiration/ViewInspiration/ViewInspiration';
@@ -36,5 +36,5 @@ export AddLink from './Links/AddLink/AddLink';
 export UpdateLink from './Links/UpdateLink/UpdateLink';
 export DeleteLink from './Links/DeleteLink/DeleteLink';
 export LinkList from './Links/LinkList/LinkList';
-export AyaArticles from './Aya/Articles';
+export AyaArticles from './Aya/AyaArticleList/AyaArticles';
 export AyaProducts from './Aya/AyaProductList/AyaProducts';
